Default to detailed input display
